Antialiasing sucks.
Because the progress bar is a canvas, the buffer becomes antialiased,
blurring the buffer. On especially short videos, you can't even tell
where the buffer ends. A temporary fix is to just set the canvas width
to it's offsetWidth (times 2 to account for DPI), and truncate `start`
and `delta` to the nearest pixel value.